interface: Add outside_write() context manager
pyout's update logic is easily thrown off if callers write to the
screen.  Provide a method that callers can use to block pyout
(optionally clearing the table), write output, and then restore the
table's content.

Output can of course come from many places that callers do not expect
or control, so this is only helpful in a subset of cases.