interface: Don't try to clear summary more than once
If a table has a summary, _write_update() clears it.  However, the
write function is called a second time if an unknown column is
encountered.  In that case, there's no summary to clear, leading to
the counts being off and lines upstream of the first row being
overwritten.

Extend the write functions with a parameter to enable distinguishing
the first and second call so that write functions can avoid any logic
that shouldn't be repeated.

Closes #128.