OPT: cache validation of styles
While troubleshooting slow "dandi download" we saw
that considerable amount of time is spent on jsonschema
validation (and logging, not considered here). See
https://github.com/dandi/dandi-cli/issues/1549

It is happening since we be providing styling per each row, which might
differ for summary line etc. And we might be re-validating over and
over again. The obvious solution, without trying to fix logic upstairs,
is to use @lru_cache.  But it could be not used as-is since style is a
dict (mutable), nested, and can have function callbacks.  Conversion of
such style to `str` is relatively fast and should be good enough for
jsonschema validation.  Unfortunately an initial attempt to just wrap
into some "Hashable" instance lead to jsonschema pukes.

So I decided to wrap/unwrap the value and use has of md5 over str
representation. Seems to work nicely and I hope it would eliminate some
elevate performance issue a little (it is not an ultimate solution
in original issue -- likely need to "slow down" records reporting).