First get content for summary and only then clear prior one
Getting summary might take time. If we wipe out prior one
first, in effect we seems to not show it for long whenever
most of the time we deal with computing the summary.
As a result we cause it to blink severily.

By first getting the summary content while prior one is still displayed,
and only then wiping it out, we seems to avoid this issue and summary
remains visible nicely without blinking.

Closes #154